+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 zixaphir
 - More mascot and theme editor fixes.
+- Fix stubs in reply hiding menu not following stubs config.
 
 ### 2.0.1 - 2013-05-08
 seaweed